rename and refactor testfiles
