+++ example_fast_concurrent_test.go
