+++ example_lite_concurrent_test.go
